slg/pyproject.toml
[project]
name = "slg"
version = "0.1.0"
description = ""
authors = [
    {name = "Ren-S07",email = "suzukiren@basicinc.jp"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "google-auth (>=2.38.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.162.0,<3.0.0)",
    "google-ads (>=25.2.0,<26.0.0)",
    "pandas (==2.1.4)",
    "numpy (>=1.23.2,<2)",
    "tabulate (>=0.9.0,<0.10.0)",
    "langgraph (>=0.3.5,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langgraph-cli[inmem] (>=0.1.55)",
    "pandasql (>=0.7.3,<0.8.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "langchain-core (>=0.3.44,<0.4.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"